<img width="1920" alt="image" src="https://github.com/user-attachments/assets/e42d558f-74d5-4e5d-aec6-dc1acae2d8e9" />

<img width="1920" alt="image" src="https://github.com/user-attachments/assets/a9c8ce7b-c7e3-4fbb-826e-fc488c20df5d" />

- [x] Highing up the perfomance by font-display  property

1. First Contentful Paint
0.8 s
2.Largest Contentful Paint
0.9 s
3.Total Blocking Time
0 ms
4.Cumulative Layout Shift
0.001
5.Speed Index
0.8 s